requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
xgboost>=1.7.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
tqdm>=4.64.0
openpyxl>=3.0.0
requests>=2.28.0

